refactor: move `get_peer_best_height` into `PeerPool`
